requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.114.0
h11==0.14.0
idna==3.8
pydantic==2.9.1
pydantic_core==2.23.3
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6

pip~=22.0.4
wheel~=0.41.3
Pillow~=10.1.0
requests~=2.31.0
urllib3~=2.0.7
certifi~=2023.7.22
PyYAML~=6.0.1
setuptools~=58.1.0
SQLAlchemy~=2.0.36
pymysql==1.0.2
python-dotenv==0.21.1